# expence-tracker/my-app/lib/schemas/user.ts
import { z } from 'zod'

// User registration schema
export const userRegistrationSchema = z.object({
  email: z.string().email('Invalid email format'),
  password: z
    .string()
    .min(8, 'Password must be at least 8 characters')
    .regex(/[A-Z]/, 'Password must contain at least one uppercase letter')
    .regex(/[0-9]/, 'Password must contain at least one number'),
})

// User login schema
export const userLoginSchema = z.object({
  email: z.string().email('Invalid email format'),
  password: z.string().min(1, 'Password is required'),
})

// Password reset request schema
export const passwordResetRequestSchema = z.object({
  email: z.string().email('Invalid email format'),
})

// Password reset schema
export const passwordResetSchema = z.object({
  token: z.string().min(1, 'Reset token is required'),
  password: z
    .string()
    .min(8, 'Password must be at least 8 characters')
    .regex(/[A-Z]/, 'Password must contain at least one uppercase letter')
    .regex(/[0-9]/, 'Password must contain at least one number'),
})

// User preferences schema
export const userPreferencesSchema = z.object({
  budgetAlertsEnabled: z.boolean(),
  dailyRemindersEnabled: z.boolean(),
  reminderTime: z
    .string()
    .regex(/^([01]\d|2[0-3]):([0-5]\d)$/, 'Time must be in HH:MM format')
    .nullable(),
  notificationMethod: z.enum(['IN_APP', 'EMAIL', 'BOTH']),
})

// Type exports
export type UserRegistration = z.infer<typeof userRegistrationSchema>
export type UserLogin = z.infer<typeof userLoginSchema>
export type PasswordResetRequest = z.infer<typeof passwordResetRequestSchema>
export type PasswordReset = z.infer<typeof passwordResetSchema>
export type UserPreferences = z.infer<typeof userPreferencesSchema>