# expence-tracker/my-app/specs/1-expense-tracker-mvp/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  description: RESTful API for the Expense Tracker MVP
  version: 1.0.0
  contact:
    name: API Support
    email: support@expensetracker.example.com
servers:
  - url: https://api.expensetracker.example.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /user/preferences:
    put:
      tags:
        - User
      summary: Update user preferences
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferences'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
    UserPreferences:
      type: object
      properties:
        budgetAlertsEnabled:
          type: boolean
        dailyRemindersEnabled:
          type: boolean
        reminderTime:
          type: string
          pattern: '^([01]\d|2[0-3]):([0-5]\d)$'
          nullable: true
          example: '09:00'
        notificationMethod:
          type: string
          enum: [IN_APP, EMAIL, BOTH]
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          nullable: true
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...